SOF-267: Moved zoomPanGesture to ModifiereExtensions, updated containerSize calculation for consistency
