+++ app/src/main/java/com/vci/vectorcamapp/imaging/presentation/components/specimen/CapturedSpecimenTile.kt
@@ -26,13 +26,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.toSize
 import coil3.compose.AsyncImage
 import coil3.request.ImageRequest
 import coil3.request.crossfade
@@ -42,14 +40,12 @@ import com.vci.vectorcamapp.core.domain.model.Specimen
 import com.vci.vectorcamapp.core.presentation.components.form.TextEntryField
 import com.vci.vectorcamapp.core.presentation.components.pill.InfoPill
 import com.vci.vectorcamapp.core.presentation.components.tile.InfoTile
-import com.vci.vectorcamapp.core.presentation.util.zoomPanGesture
 import com.vci.vectorcamapp.imaging.presentation.components.camera.BoundingBoxOverlay
 import com.vci.vectorcamapp.ui.extensions.colors
 import com.vci.vectorcamapp.ui.extensions.dimensions
+import com.vci.vectorcamapp.ui.extensions.zoomPanGesture
 import java.text.SimpleDateFormat
 import java.util.Locale
-import net.engawapg.lib.zoomable.rememberZoomState
-import net.engawapg.lib.zoomable.zoomable
 
 @Composable
 fun CapturedSpecimenTile(
@@ -73,7 +69,8 @@ fun CapturedSpecimenTile(
         ) {
             val containerSize = IntSize(
                 width = with(density) { maxWidth.roundToPx() },
-                height = with(density) { maxHeight.roundToPx() })
+                height = with(density) { maxHeight.roundToPx() }
+            )
 
             Box(
                 modifier = Modifier
